optiver_app/requirements.txt
boto3==1.34.86
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
fastapi==0.110.1
uvicorn[standard]==0.29.0
pandas==2.2.2
uvicorn==0.29.0
requests==2.31.0
pydantic==2.7.0
gunicorn==22.0.0